mcp_server/requirements.txt
fastapi
uvicorn
google-cloud-bigquery
pandas
pyarrow
db-dtypes